Move some namespaces on reflection
