+++ src/Bundle/DependencyInjection/SyliusResourceExtension.php
@@ -21,12 +21,12 @@ use Sylius\Bundle\ResourceBundle\DependencyInjection\Driver\DriverProvider;
 use Sylius\Bundle\ResourceBundle\Form\Type\DefaultResourceType;
 use Sylius\Bundle\ResourceBundle\SyliusResourceBundle;
 use Sylius\Component\Resource\Factory\FactoryInterface as LegacyFactoryInterface;
-use Sylius\Component\Resource\Reflection\ClassReflection;
 use Sylius\Resource\Factory\Factory;
 use Sylius\Resource\Factory\FactoryInterface;
 use Sylius\Resource\Metadata\AsResource;
 use Sylius\Resource\Metadata\Metadata;
 use Sylius\Resource\Metadata\ResourceMetadata;
+use Sylius\Resource\Reflection\ClassReflection;
 use Sylius\Resource\State\ProcessorInterface;
 use Sylius\Resource\State\ProviderInterface;
 use Sylius\Resource\State\ResponderInterface;
