+++ src/Bundle/Routing/RoutesAttributesLoader.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\Routing;
 
-use Sylius\Component\Resource\Reflection\ClassReflection;
+use Sylius\Resource\Reflection\ClassReflection;
 use Symfony\Bundle\FrameworkBundle\Routing\RouteLoaderInterface;
 use Symfony\Component\Routing\RouteCollection;
