+++ src/Component/Reflection/ClassReflection.php
@@ -13,58 +13,11 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Resource\Reflection;
 
-use Symfony\Component\Finder\Finder;
+class_exists(\Sylius\Resource\Reflection\ClassReflection::class);
 
-final class ClassReflection
-{
-    public static function getResourcesByPaths(array $paths): iterable
+if (false) {
+    final class ClassReflection extends \Sylius\Resource\Reflection\ClassReflection
     {
-        foreach ($paths as $resourceDirectory) {
-            $resources = self::getResourcesByPath($resourceDirectory);
-
-            foreach ($resources as $className) {
-                yield $className;
-            }
-        }
-    }
-
-    public static function getResourcesByPath(string $path): iterable
-    {
-        $finder = new Finder();
-        $finder->files()->in($path)->name('*.php')->sortByName(true);
-
-        foreach ($finder as $file) {
-            $fileContent = (string) file_get_contents((string) $file->getRealPath());
-
-            preg_match('/namespace (.+);/', $fileContent, $matches);
-
-            $namespace = $matches[1] ?? null;
-
-            if (!preg_match('/class +([^{ ]+)/', $fileContent, $matches)) {
-                // no class found
-                continue;
-            }
-
-            $className = trim($matches[1]);
-
-            if (null !== $namespace) {
-                yield $namespace . '\\' . $className;
-            } else {
-                yield $className;
-            }
-        }
-    }
-
-    /**
-     * @psalm-param class-string $className
-     *
-     * @return \ReflectionAttribute[]
-     */
-    public static function getClassAttributes(string $className, ?string $attributeName = null): array
-    {
-        $reflectionClass = new \ReflectionClass($className);
-
-        /** @psalm-suppress ArgumentTypeCoercion */
-        return $reflectionClass->getAttributes($attributeName);
     }
 }
+
