+++ src/Component/src/Symfony/Request/State/Provider.php
@@ -15,12 +15,12 @@ namespace Sylius\Resource\Symfony\Request\State;
 
 use Pagerfanta\Pagerfanta;
 use Psr\Container\ContainerInterface;
-use Sylius\Component\Resource\Reflection\CallableReflection;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Option\RequestOption;
 use Sylius\Resource\Metadata\BulkOperationInterface;
 use Sylius\Resource\Metadata\CollectionOperationInterface;
 use Sylius\Resource\Metadata\Operation;
+use Sylius\Resource\Reflection\CallableReflection;
 use Sylius\Resource\State\ProviderInterface;
 use Sylius\Resource\Symfony\ExpressionLanguage\ArgumentParserInterface;
 use Sylius\Resource\Symfony\Request\RepositoryArgumentResolver;
